requirements.txt
requests~=2.32.3
xmltodict~=0.13.0
pyyaml~=6.0.1
flask~=3.0.3
pymongo~=4.7.3